Make tile culling implementation more flexible
The previous implementation is slightly more efficient but may cause issues down the line, e.g., when adding super-sampling. Instead of considering the full tile, the old implementation skipped the outer boundary of the tile, e.g., a 15x15 region instead of the full 16x16 tile. This is possible with the default 3DGS rendering because splats are only evaluated at the center of each pixel.